Changed influence per minute to a float. Created a file for constants, tweaked IPM values

+++ database/migrations/2016_04_13_085412_create_soldiers_table.php
@@ -16,7 +16,7 @@ class CreateSoldiersTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned()->index();
             $table->string('name');
-            $table->integer('influence_per_minute')->unsigned();
+            $table->float('influence_per_minute')->unsigned();
             $table->timestamps();
         });
     }
